Upgrade Connection.Weight and Eligibility from int32 to int64
Preserves more precision when quantizing float64 training weights
for integer inference. All activation math remains int32.

Changes:
- Connection.Weight, Eligibility: int32 -> int64
- PendingStimulation.Weight: int32 -> int64
- Network.Connect, Network.Stimulate: accept int64
- ClampAdd: operates on int64 (for weights); ClampAdd32 added for activation
- MaxWeight/MinWeight: now int64 constants
- syncWeightsToNetwork: quantizes to int64
- All learning rules updated (stdp, rstdp, predictive, arbiter, perturbation)
- All benchmarks and tests updated
- Iris surrogate test: fixed unseeded RNG for deterministic runs

Verified: go vet clean, all non-MNIST tests pass, MNIST verification
in progress (epoch-by-epoch identical to int32 baseline through 15/30).

+++ benchmark/cmd/xor/main.go
@@ -150,14 +150,14 @@ func runWithReward(rule bio.LearningRule, name string, cfg xor.NetworkConfig) *b
 // This is the key mechanism for discovering inhibitory connections.
 func applyAsymmetricUpdate(net *bio.Network, layout xor.Layout, reward int32, window uint32) {
 	now := net.Counter
-	maxW := int32(5000) // cap weight magnitude
+	maxW := int64(5000) // cap weight magnitude
 
 	isRecentlyActive := func(idx uint32) bool {
 		n := &net.Neurons[idx]
 		return n.LastFired > 0 && now-n.LastFired < window
 	}
 
-	clampWeight := func(w int32) int32 {
+	clampWeight := func(w int64) int64 {
 		if w > maxW {
 			return maxW
 		}
@@ -178,9 +178,9 @@ func applyAsymmetricUpdate(net *bio.Network, layout xor.Layout, reward int32, wi
 			tgtActive := isRecentlyActive(conn.Target)
 
 			if srcActive && tgtActive {
-				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, reward/4))
+				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, int64(reward)/4))
 			} else if srcActive && !tgtActive {
-				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, -reward/8))
+				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, -int64(reward)/8))
 			}
 		}
 	}
@@ -196,9 +196,9 @@ func applyAsymmetricUpdate(net *bio.Network, layout xor.Layout, reward int32, wi
 			tgtActive := isRecentlyActive(conn.Target)
 
 			if srcActive && tgtActive {
-				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, reward/4))
+				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, int64(reward)/4))
 			} else if srcActive && !tgtActive {
-				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, -reward/8))
+				conn.Weight = clampWeight(bio.ClampAdd(conn.Weight, -int64(reward)/8))
 			}
 		}
 	}
