+++ benchmark/mnist/mnist_deep_test.go
@@ -57,7 +57,7 @@ func TestMNISTDeep(t *testing.T) {
 					continue
 				}
 				wf := (rng.Float64()*2.0 - 1.0) * initWeightMax
-				w := int32(math.Round(wf * intScale))
+				w := int64(math.Round(wf * intScale))
 				if w == 0 {
 					w = 1
 				}
